renamed handybars_attribute crate to handybars_macro crate and introduces  feature
