+++ .github/workflows/rust.yml
@@ -17,6 +17,6 @@ jobs:
     steps:
     - uses: actions/checkout@v4
     - name: Build
-      run: cargo build --verbose
+      run: cargo build --verbose --workspace --all-features
     - name: Run tests
-      run: cargo test --verbose
+      run: cargo test --verbose --workspace --all-features
