+++ Cargo.toml
@@ -7,12 +7,17 @@ authors = ["0x00002a"]
 readme = "README.md"
 description = "Minimal template parsing and expansion"
 
-[workspace]
-members = [ "." , "attribute" ]
+[features]
+macros = [ "dep:handybars_macros" ]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[package.metadata."docs.rs"]
+all-features = true
+
+[workspace]
+members = [ "." , "macros" ]
 
 [dependencies]
+handybars_macros = { path = "macros", optional = true }
 
 [dev-dependencies]
 proptest = "1.0.0"
