+++ README.md
@@ -12,7 +12,7 @@ It has no dependencies and is designed to have a very simple API.
 A simple [companion attribute macro](attribute/README.md) allows you to automatically turn enums and
 structs into Handybar [Value](https://docs.rs/handybars/latest/handybars/enum.Value.html) variants.
 
-## Usage
+## Basic Usage
 
 ```rust
 use handybars::{Context, Variable};
@@ -40,4 +40,134 @@ Note that objects cannot be directly expanded:
 use handybars::{Context, Variable, Object, Error};
 let ctx = Context::new().with_define("world".parse().unwrap(), Object::new().with_property("a", "p1"));
 assert_eq!(ctx.render("{{world}}"), Err(Error::TriedToExpandObject(Variable::single("world"))));
+```
+## Macros Usage
+Make sure to include the `macros` feature for the Handybar dependenc in `Cargo.toml`:
+```toml
+[dependencies]
+handybars = { version = "0.2", features = [ "macros" ] }
+```
+Import the handybars and the macro in your rust code:
+```rust
+# #[cfg(feature = "macros")]
+use handybars::handybars_value;
+```
+Use simple enums as [Value::String](https://docs.rs/handybars/latest/handybars/enum.Value.html) variants:
+```rust
+# #[cfg(feature = "macros")]
+# use handybars::handybars_value;
+# #[cfg(feature = "macros")]
+#[handybars_value]
+enum SimpleEnumProp {
+    A,
+    B,
+}
+```
+Use structs as [Value::Object](https://docs.rs/handybars/latest/handybars/enum.Value.html) variants::
+```rust
+# #[cfg(feature = "macros")]
+# use handybars::handybars_value;
+# #[handybars_value]
+# #[cfg(feature = "macros")]
+# enum SimpleEnumProp {
+#     A,
+#     B,
+# }
+# #[cfg(feature = "macros")]
+#[handybars_value]
+struct StructVal<'a> {
+    field_1: u16,
+    field_2: String,
+    field_3: &'a str,
+    field_4: SimpleEnumProp,
+}
+```
+Combine enums and structs into more complex objects:
+```rust
+# #[cfg(feature = "macros")]
+# use handybars::handybars_value;
+# #[cfg(feature = "macros")]
+# #[handybars_value]
+# enum SimpleEnumProp {
+#     A,
+#     B,
+# }
+# #[cfg(feature = "macros")]
+# #[handybars_value]
+# struct StructVal<'a> {
+#     field_1: u16,
+#     field_2: String,
+#     field_3: &'a str,
+#     field_4: SimpleEnumProp,
+# }
+# #[cfg(feature = "macros")]
+#[handybars_value]
+struct TestObject<'a> {
+    prop_0: String,
+    prop_1: u64,
+    prop_2: &'a str,
+    prop_3: StructVal<'a>,
+    prop_4: SimpleEnumProp,
+}
+```
+Example on using the above enums and structs:
+```rust
+# #[cfg(feature = "macros")]
+# use handybars::{ Context, Variable, handybars_value};
+# #[cfg(feature = "macros")]
+# #[handybars_value]
+# enum SimpleEnumProp {
+#     A,
+#     B,
+# }
+# #[cfg(feature = "macros")]
+# #[handybars_value]
+# struct StructVal<'a> {
+#     field_1: u16,
+#     field_2: String,
+#     field_3: &'a str,
+#     field_4: SimpleEnumProp,
+# }
+# #[cfg(feature = "macros")]
+# #[handybars_value]
+# struct TestObject<'a> {
+#     prop_0: String,
+#     prop_1: u64,
+#     prop_2: &'a str,
+#     prop_3: StructVal<'a>,
+#     prop_4: SimpleEnumProp,
+# }
+# #[cfg(feature = "macros")]
+let v = TestObject {
+    prop_0: "p0_val".to_owned(),
+    prop_1: 1,
+    prop_2: "p2_val",
+    prop_3: StructVal {
+        field_1: 30,
+        field_2: "f32_val".to_owned(),
+        field_3: "f33_val",
+        field_4: SimpleEnumProp::A,
+    },
+    prop_4: SimpleEnumProp::B,
+};
+# #[cfg(feature = "macros")]
+let c = Context::new().with_define(Variable::single("obj"), v);
+# #[cfg(feature = "macros")]
+assert_eq!("1", c.render("{{ obj.prop_1 }}").unwrap());
+# #[cfg(feature = "macros")]
+assert_eq!("A", c.render("{{ obj.prop_3.field_4 }}").unwrap());
+# #[cfg(feature = "macros")]
+assert_eq!("f33_val", c.render("{{ obj.prop_3.field_3 }}").unwrap());
+```
+The running code for the above can be found as a [macro test case](tests/handybars_macro.rs).
+
+> [!WARNING]
+> Enums with variant values are currently **not supported**. Enum with variants like the following will not compile:
+```compile_fail
+#[handybar_value]
+enum ComplexEnumProp<'a> {
+    Var1(SimpleEnumProp),
+    Var2(String),
+    Var3(StructVal<'a>),
+}
 ```
