+++ macros/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "handybars_attribute"
+name = "handybars_macros"
 version = "0.2.0"
 edition = "2021"
 authors = [ "chrsoo" ]
@@ -12,4 +12,3 @@ proc-macro = true
 [dependencies]
 syn = {version = "2.0", features = ["full"]}
 quote = "1.0"
-handybars = { path = ".."}
